Async Implemented in Employee Controller

+++ Web Development/HumanResources/HumanResources.Web/Controllers/EmployeeController.cs
@@ -13,27 +13,27 @@ public class EmployeeController : Controller
 	{
 		db = _db;
 	}
-	public IActionResult Index()
+	public async Task<IActionResult> Index()
 	{
-		var employees = db.Employees.Include(e => e.Department).Include(e=>e.Designation).ToList();
+		var employees = await db.Employees.Include(e => e.Department).Include(e=>e.Designation).ToListAsync();
 
 		return View(employees);
 	}
 
-	public IActionResult Add()
+	public async  Task<IActionResult> Add()
 	{
-		var departments = db.Departments.ToList();
-		var departmentList = departments.Select(x => new SelectListItem { Text = x.Name, Value = x.Id.ToString() });
-		ViewData["departments"] = departmentList;
+		//var departments = db.Departments.ToList();
+		//var departmentList = departments.Select(x => new SelectListItem { Text = x.Name, Value = x.Id.ToString() });
+		//ViewData["departments"] = departmentList;
 
-		var designations = db.Designations.ToList();
-		var designationList = designations.Select(x=> new SelectListItem { Text=x.Name, Value=x.Id.ToString() });
-		ViewData["designations"] = designationList;
+		//var designations = db.Designations.ToList();
+		//var designationList = designations.Select(x=> new SelectListItem { Text=x.Name, Value=x.Id.ToString() });
+		//ViewData["designations"] = designationList;
 
-		//var departments = db.Departments.Select(x => new SelectListItem { Text = x.Name, Value = x.Name }).ToList();
-		//ViewData["departments"] = departments;
-		//var designations = db.Designations.Select(x => new SelectListItem { Text = x.Name, Value = x.Name }).ToList();
-		//ViewData["designations"] = designations;
+		var departments =await db.Departments.Select(x => new SelectListItem { Text = x.Name, Value = x.Id.ToString() }).ToListAsync();
+		ViewData["departments"] = departments;
+		var designations = await db.Designations.Select(x => new SelectListItem { Text = x.Name, Value = x.Id.ToString() }).ToListAsync();
+		ViewData["designations"] = designations;
 		//return View();
 
 		return View();
@@ -64,15 +64,20 @@ public class EmployeeController : Controller
 		return relativePath;
 	}
 
-	public IActionResult Edit(int id)
+	public async Task<IActionResult> Edit(int id)
 	{
-		var department = db.Departments.ToList();
-		var departmentLis = department.Select(x => new SelectListItem { Text = x.Name, Value = x.Id.ToString() });
-		ViewData["department"] = departmentLis;
+		//var department = db.Departments.ToList();
+		//var departmentLis = department.Select(x => new SelectListItem { Text = x.Name, Value = x.Id.ToString() });
+		//ViewData["department"] = departmentLis;
 
-		var designation = db.Designations.ToList();
-		var designationLis = designation.Select(x => new SelectListItem { Text = x.Name, Value = x.Id.ToString() });
-		ViewData["designation"] = designationLis;
+		//var designation = db.Designations.ToList();
+		//var designationLis = designation.Select(x => new SelectListItem { Text = x.Name, Value = x.Id.ToString() });
+		//ViewData["designation"] = designationLis;
+
+		var department = await db.Departments.Select(x => new SelectListItem { Text = x.Name, Value = x.Id.ToString() }).ToListAsync();
+		ViewData["department"] = department;
+		var designation = await db.Designations.Select(x => new SelectListItem { Text = x.Name, Value = x.Id.ToString() }).ToListAsync();
+		ViewData["designation"] = designation;
 		var employee = db.Employees.Find(id);
 		return View(employee);
 	}
@@ -91,15 +96,27 @@ public class EmployeeController : Controller
 
 	public IActionResult Delete(int id)
 	{
+		if (id == null || db.Departments == null)
+		{
+			return NotFound();
+		}
 		var employee = db.Employees.Find(id);
 		return View(employee);
 	}
 
 	[HttpPost]
-	public IActionResult Delete(Employee employee)
+	public async Task<IActionResult> Delete(Employee employee)
 	{
-		db.Employees.Remove(employee);
-		db.SaveChanges();
+		if (db.Employees== null)
+		{
+			return Problem("Entity set 'HRDbContext.Employees'  is null.");
+		}
+		if (employee != null)
+		{
+			db.Employees.Remove(employee);
+		}
+
+		await db.SaveChangesAsync();
 
 		return RedirectToAction("Index");
 	}
